[red-knot] Extract `red_knot_python_semantic` crate
