+++ crates/red_knot_python_semantic/src/db.rs
@@ -6,11 +6,10 @@ use crate::module::resolver::{
     file_to_module, internal::ModuleNameIngredient, internal::ModuleResolverSearchPaths,
     resolve_module_query,
 };
-use crate::red_knot::semantic_index::symbol::{
-    public_symbols_map, scopes_map, PublicSymbolId, ScopeId,
-};
-use crate::red_knot::semantic_index::{root_scope, semantic_index, symbol_table};
-use crate::red_knot::types::{infer_types, public_symbol_ty};
+
+use crate::semantic_index::symbol::{public_symbols_map, scopes_map, PublicSymbolId, ScopeId};
+use crate::semantic_index::{root_scope, semantic_index, symbol_table};
+use crate::types::{infer_types, public_symbol_ty};
 
 #[salsa::jar(db=Db)]
 pub struct Jar(
