+++ crates/red_knot_python_semantic/src/module/resolver.rs
