+++ crates/red_knot_python_semantic/src/node_key.rs
