+++ crates/ruff_python_semantic/src/lib.rs
@@ -2,17 +2,10 @@ pub mod analyze;
 mod binding;
 mod branches;
 mod context;
-#[cfg(feature = "red_knot")]
-mod db;
 mod definition;
 mod globals;
 mod model;
-#[cfg(feature = "red_knot")]
-pub mod module;
-pub mod name;
 mod nodes;
-#[cfg(feature = "red_knot")]
-pub mod red_knot;
 mod reference;
 mod scope;
 mod star_import;
@@ -27,6 +20,3 @@ pub use nodes::*;
 pub use reference::*;
 pub use scope::*;
 pub use star_import::*;
-
-#[cfg(feature = "red_knot")]
-pub use db::{Db, Jar};
